Move UUIDType to correct package
